adds entity radix tree
